Create new NavbarSearch component
