+++ frontend/src/components/NavbarSearch.module.css
@@ -1,0 +1,11 @@
+.search_bar {
+  width: 40vw;
+}
+
+.full_height {
+  height: 100%;
+}
+
+.half_height {
+  height: 50%;
+}
